More application services

+++ backend/backend/Product/InfrastructureLayer/ProductRepository.cs
@@ -14,8 +14,14 @@ public class ProductRepository(ProductDbContext context) : IProductRepository
         await context.SaveChangesAsync();
     }
 
-    public async Task DeleteProductAsync(DomainModels.Product product)
+    public async Task DeleteProductAsync(long id)
     {
+        var product = await context.Products.FindAsync(id);
+        if (product == null)
+        {
+            throw new KeyNotFoundException($"Product with ID {id} not found.");
+        }
+
         context.Products.Remove(product);
         await context.SaveChangesAsync();
     }
